Implement findNftsByMintList
