+++ src/modules/nfts/index.ts
@@ -12,13 +12,14 @@ import * as h from './operationHandlers';
 
 export const nftPlugin = (): MetaplexPlugin => ({
   install(metaplex: Metaplex) {
-    metaplex.register(o.UploadMetadataOperation, h.UploadMetadataOperationHandler);
-    metaplex.register(o.PlanUploadMetadataOperation, h.PlanUploadMetadataOperationHandler);
     metaplex.register(o.CreateNftOperation, h.CreateNftOperationHandler);
-    metaplex.register(o.UpdateNftOperation, h.UpdateNftOperationHandler);
     metaplex.register(o.FindNftByMintOperation, h.FindNftByMintOperationHandler);
     metaplex.register(o.FindNftsByCandyMachineOperation, h.FindNftsByCandyMachineOperationHandler);
+    metaplex.register(o.FindNftsByMintListOperation, h.FindNftsByMintListOperationHandler);
     metaplex.register(o.FindNftsByOwnerOperation, h.FindNftsByOwnerUsingGpasOperationHandler);
+    metaplex.register(o.PlanUploadMetadataOperation, h.PlanUploadMetadataOperationHandler);
+    metaplex.register(o.UpdateNftOperation, h.UpdateNftOperationHandler);
+    metaplex.register(o.UploadMetadataOperation, h.UploadMetadataOperationHandler);
 
     metaplex.nfts = function () {
       return new NftClient(this);
