+++ src/modules/nfts/operationHandlers/index.ts
@@ -1,6 +1,7 @@
 export * from './CreateNftOperationHandler';
 export * from './FindNftByMintOperationHandler';
 export * from './FindNftsByCandyMachineOperationHandler';
+export * from './FindNftsByMintListOperationHandler';
 export * from './FindNftsByOwnerUsingGpasOperationHandler';
 export * from './PlanUploadMetadataOperationHandler';
 export * from './UpdateNftOperationHandler';
